+++ insight-brain-db/src/main/resources/db/insight_brain_ods/schema.sql
@@ -815,7 +815,7 @@ CREATE TABLE source_control_pull_request_comment
   component_hash varchar(20),
   pathname varchar(1000) NULL,
   pull_request_id int NOT NULL,
-  pull_request_comment_id int NOT NULL,
+  pull_request_comment_id bigint NOT NULL,
   pull_request_comment_version int default NULL,
   source_policy_evaluation_id varchar(50) NOT NULL,
   target_policy_evaluation_id varchar(50) NOT NULL,
